Crate namada_vp

This crate contains the trait for native validity predicates with its various context types and host functions implementation.

Re-exports

pub use namada_state as state;

Modules

native_vp

Native validity predicate interface associated with internal accounts such as the PoS and IBC modules.

vp_host_fns

Host functions for VPs used for both native and WASM VPs.

Structs

Event

Custom events that can be queried from Tendermint using a websocket client

EventType

ABCI event type.

Hash

A hash, typically a sha-2 hash of a tx

Enums

Address

An account’s address

Traits

VpEnv

Validity predicate’s environment is available for native VPs and WASM VPs