namada_vp
Module native_vp

Native validity predicate interface associated with internal accounts such as the PoS and IBC modules.

Structs

Ctx

A validity predicate’s host context.

CtxPostStorageRead

Read access to the posterior storage (state after tx execution) via StorageRead.

CtxPreStorageRead

Read access to the prior storage (state before tx execution) via StorageRead.

Enums

Error

Traits

NativeVp

A native VP module should implement its validation logic using this trait.

ResultExt

Result extension to easily wrap custom errors into Error.

StorageReader

A convenience trait for reading and automatically deserializing a value from storage

VpEvaluator

A Validity predicate runner for calls from the host env vp_eval function.

Type Aliases

Result

Result of a storage API call.