namada_vp::native_vp
Trait StorageReader

pub trait StorageReader {
    // Required methods
    fn read_pre_value<T: BorshDeserialize>(
        &self,
        key: &Key,
    ) -> Result<Option<T>>;
    fn read_post_value<T: BorshDeserialize>(
        &self,
        key: &Key,
    ) -> Result<Option<T>>;

    // Provided methods
    fn must_read_pre_value<T: BorshDeserialize>(&self, key: &Key) -> Result<T> { ... }
    fn must_read_post_value<T: BorshDeserialize>(&self, key: &Key) -> Result<T> { ... }
}

A convenience trait for reading and automatically deserializing a value from storage

Required Methods

fn read_pre_value<T: BorshDeserialize>(&self, key: &Key) -> Result<Option<T>>

Storage read prior state (before tx execution). It will try to read from the storage.

fn read_post_value<T: BorshDeserialize>(&self, key: &Key) -> Result<Option<T>>

Storage read posterior state (after tx execution). It will try to read from the write log first and if no entry found then from the storage.

Provided Methods

fn must_read_pre_value<T: BorshDeserialize>(&self, key: &Key) -> Result<T>

Calls read_pre_value, and returns an error on Ok(None).

fn must_read_post_value<T: BorshDeserialize>(&self, key: &Key) -> Result<T>

Calls read_post_value, and returns an error on Ok(None).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.