namada_vp
Module vp_host_fns

Host functions for VPs used for both native and WASM VPs.

Enums

Error

RuntimeError

These runtime errors will abort VP execution immediately

Functions

add_gas

Add a gas cost incured in a validity predicate

get_block_epoch

Getting the block epoch. The epoch is that of the block to which the current transaction is being applied.

get_block_header

Getting the block header.

get_block_height

Getting the block height. The height is that of the block to which the current transaction is being applied.

get_chain_id

Getting the chain ID.

get_events

Query events emitted by the current transaction.

get_native_token

Getting the native token’s address.

get_pred_epochs

Given the information about predecessor block epochs

get_tx_code_hash

Getting the block hash. The height is that of the block to which the current transaction is being applied.

get_tx_index

Getting the block epoch. The epoch is that of the block to which the current transaction is being applied.

has_key_post

Storage has_key in posterior state (after tx execution). It will try to check the write log first and if no entry found then the storage.

has_key_pre

Storage has_key in prior state (before tx execution). It will try to read from the storage.

iter_next

Get the next item in a storage prefix iterator (pre or post).

iter_prefix_post

Storage prefix iterator for posterior state (after tx execution), ordered by storage keys. It will try to get an iterator from the storage.

iter_prefix_pre

Storage prefix iterator for prior state (before tx execution), ordered by storage keys. It will try to get an iterator from the storage.

read_post

Storage read posterior state (after tx execution). It will try to read from the write log first and if no entry found then from the storage.

read_pre

Storage read prior state (before tx execution). It will try to read from the storage.

read_temp

Storage read temporary state (after tx execution). It will try to read from only the write log.

Type Aliases

EnvResult

VP environment function result

Result

Result of a storage API call.